+++ examples/tutorials/hello-cube/app.ts
@@ -5,7 +5,7 @@
 import type {NumberArray, ShaderUniformType} from '@luma.gl/core';
 import {glsl, Texture, UniformStore} from '@luma.gl/core';
 import type {AnimationProps} from '@luma.gl/engine';
-import {AnimationLoopTemplate, Model, CubeGeometry, loadImage} from '@luma.gl/engine';
+import {AnimationLoopTemplate, Model, CubeGeometry, loadImage, AsyncTexture} from '@luma.gl/engine';
 import {Matrix4} from '@math.gl/core';
 
 const INFO_HTML = `\
@@ -120,7 +120,7 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
   constructor({device}: AnimationProps) {
     super();
 
-    const texture = device.createTexture({
+    const texture = new AsyncTexture(device, {
       usage: Texture.TEXTURE & Texture.COPY_DST,
       data: loadImage('vis-logo.png'),
       mipmaps: true,