+++ examples/tutorials/lighting/app-old.ts
@@ -1,5 +1,6 @@
 import {glsl, NumberArray, UniformStore, ShaderUniformType} from '@luma.gl/core';
-import {AnimationLoopTemplate, AnimationProps, Model, CubeGeometry} from '@luma.gl/engine';
+import type {AnimationProps} from '@luma.gl/engine';
+import {AnimationLoopTemplate, Model, CubeGeometry, loadImage, AsyncTexture} from '@luma.gl/engine';
 import {phongMaterial, lighting} from '@luma.gl/shadertools';
 import {Matrix4} from '@math.gl/core';
 
@@ -124,7 +125,7 @@ export default class AppAnimationLoopTemplate extends AnimationLoopTemplate {
       }
     });
 
-    const texture = device.createTexture({data: 'vis-logo.png'});
+    const texture = new AsyncTexture(device, {data: loadImage('vis-logo.png')});
 
     this.model = new Model(device, {
       vs,